chore: rename basic model to simple pytorch model and extract files
