+++ src/spira_training/shared/adapters/pytorch/model_trainer/implementations/pytorch_model/mish.py
